Основна робота/requirements.txt
Flask
Flask-WTF
Flask-Login
psycopg2-binary
Werkzeug
WTForms